+++ src/main/java/cn/haitaoss/tinyioc/beans/lifecycle/DisposableBean.java
@@ -1,0 +1,11 @@
+package cn.haitaoss.tinyioc.beans.lifecycle;
+
+/**
+ * @author haitao.chen
+ * @email haitaoss@aliyun.com
+ * @date 2021-04-22 21:36
+ *
+ */
+public interface DisposableBean {
+    void destroy();
+}